chore: fix clippy -D warnings in nova-syntax/nova-hir
CI runs clippy with -D warnings; recent code paths triggered a set of new lints.

- nova-syntax: remove duplicate match arm, prefer is_empty/is_none_or/find, and simplify iterator usage
- nova-memory/nova-test-utils/xtask: apply small clippy-driven cleanups
- nova-hir: derive Default where applicable and remove redundant #[must_use]

Validated with:
- bash scripts/cargo_agent.sh clippy -p nova-syntax --all-targets --all-features -- -D warnings
- bash scripts/cargo_agent.sh clippy -p nova-hir --all-targets --all-features -- -D warnings
- bash scripts/cargo_agent.sh test -p nova-syntax --tests
- bash scripts/cargo_agent.sh test -p nova-hir --tests

+++ crates/nova-memory/tests/suite/cgroup.rs
@@ -101,7 +101,7 @@ fn interprets_rlimit_as_infinity_or_extremely_large_as_unlimited() {
 #[test]
 fn interprets_libc_rlim_infinity_as_unlimited() {
     assert_eq!(
-        interpret_rlimit_as_bytes(libc::RLIM_INFINITY as u64, libc::RLIM_INFINITY as u64),
+        interpret_rlimit_as_bytes(libc::RLIM_INFINITY, libc::RLIM_INFINITY),
         None
     );
 }
